Add files via upload
An Information Visualization Experiment to propose a hypothesis and perform a pilot experiment.

The experiment investigated the importance of using grid-lines when representing a data visualization in which some trends exist, that are depicted using one or more trend-line(s). The core aspects related to the comprehension of trend-lines and the importance of grid-lines was discussed. A reader’s ability to conceive the angle of a trend-line is limited. This can be observed especially while comparing charts whose trend-lines have a very minute slope difference. The minute difference may be difficult to interpret but if grid-lines are added, the reader will have a point of reference and can make more accurate interpretations of the data visualization.